[feat] 정답체크 List형태로 Body받게 변경 + 리턴 타입 int 로 변경 (정답시 1/ 오답시 0)

See merge request kofia/etest-sap-backend!53

+++ eTest/src/main/java/com/tmax/eTest/Contents/service/ProblemServices.java
@@ -51,7 +51,7 @@ public class ProblemServices {
 	
 	private final Logger logger = LoggerFactory.getLogger(this.getClass());
 	
-	public ProblemDTO getProblem(long problemID) throws Exception{
+	public ProblemDTO getProblem(Integer problemID) throws Exception{
 		ProblemDTO output;
 		Optional<Problem> problemOpt = problemRepo.findById(problemID);
